backend/requirements.txt
fastapi
uvicorn
pdfplumber
sentence-transformers
faiss-cpu
transformers
feedparser
apscheduler
python-dotenv
requests
nltk
torch
numpy